factor out varToPath
